feat: Add basic HTTP server functionality
This commit adds the basic functionality of an HTTP server using Go. It includes a simple handler function that responds with "Hello, Go API!" for any incoming request to the root ("/") endpoint. The server listens on port 8080.

The purpose of this commit is to provide a starting point for building a Go API.